YT-4130: Add join limit test

+++ yt/ytlib/query_client/cg_types.h
@@ -80,7 +80,7 @@ class TypeBuilder<TJoinLookup*, Cross>
 { };
 
 template <bool Cross>
-class TypeBuilder<std::vector<std::pair<TRow, int>>*, Cross>
+class TypeBuilder<std::vector<std::pair<TRow, i64>>*, Cross>
     : public TypeBuilder<void*, Cross>
 { };;
